app/requirements/main.txt
Django==6.0.0
django-ninja==1.5.3
black==23.9.1
isort==5.12.0
flake8==6.1.0
pre-commit==3.5.0
django-cors-headers==4.6.0
pika==1.3.2
uvicorn==0.40.0
psycopg[binary]==3.2.3
whitenoise==6.11.0
pymemcache==4.0.0
django-ninja-jwt==5.4.4
django-ninja-extra==0.31.0
